refactor: Milestone 클래스 Builder 어노테이션 추가

+++ back/src/main/java/codesquad/issuetracker/milestone/MilestoneRepository.java
@@ -2,5 +2,5 @@ package codesquad.issuetracker.milestone;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface MilestoneRepository extends CrudRepository<Milestone, Integer> {
+public interface MilestoneRepository extends CrudRepository<Milestone, Long> {
 }
